Refactor: Improve steps tracking and data handling
This commit introduces several improvements to the steps tracking feature, including enhanced UI, better data aggregation, and more robust state management.

### Key Changes:

-   **Steps Chart Enhancements:**
    -   **`PlotSteps.kt`:** Added an annotation to the steps chart that displays the exact step count when a data point is tapped.
    -   **`strings.xml`:** Added new string resources for the chart annotation (`steps_chart_annotation_steps_value`) and for accessibility in the "Add Steps" dialog (`steps_add_dialog_content_description_date`).

-   **Improved Data Aggregation and Type Safety:**
    -   **`StepsOverviewViewModel.kt`:**
        -   Refactored the data aggregation logic to handle `Double` for average step counts, allowing for more precise calculations (e.g., `123.45` steps on average) instead of rounding to an integer.
        -   The `StepsOverview` data class now uses a `Double` for `steps` instead of `UInt`.
        -   The `onSave` event in the dialog now passes an `Instant` instead of a `LocalDate` for more precise timestamping.
        -   Added error handling to the `addSteps` function to catch potential issues when saving data.
    -   **`BeverageOverviewViewModel.kt`, `SleepOverviewViewModel.kt`:** Updated `maxOfOrNull` calculations to correctly handle `Double` values and round them to `UInt` when determining the maximum value for chart scaling, preventing potential crashes.

-   **UI and State Management Cleanup:**
    -   **`AddStepsDialog.kt`:**
        -   The `onSave` callback now provides an `Instant` instead of a `LocalDate`.
        -   Added a content description to the date picker icon for better accessibility.
    -   **`AddSleepDialog.kt`:** Ensured `viewModel.dismissDialog()` is called on both confirm and dismiss actions to properly clear the dialog state.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/steps/StepsOverview.kt
@@ -5,5 +5,5 @@ import org.darthacheron.fitbe.health.components.Overview
 
 data class StepsOverview(
     override val date: LocalDate,
-    val steps: UInt
+    val steps: Double
 ) : Overview
